blog/requirements.txt
asgiref==3.2.7
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-filter==2.3.0
django-js-asset==1.2.2
djangorestframework==3.11.0
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
